Removed all the FIXME and many XXX references in the code.
Added a few missing copyright notices.
Added a #pragma once to the qlockfile.h header.
Replaced a couple of bare pointers with controlled_vars pointers.
Removed many static_cast<>() from the content plugin as they were rendered unnecessary.
Added a .submit() in the form.js script to prevent a double submit of any one form.
Fixed a type definition in the form.js.
Added a couple of casts to JavaScript code as pointed out by Google Closure.
Added tests to determine what causes the user to be logged out once in a while: as expected the random number is the problem, it happens because we send two requests from the browser before we get the reply from the first request.

+++ snapwebsites/lib/compression.cpp
@@ -509,7 +509,7 @@ public:
 
 void archiver_t::file_t::set_type(type_t type)
 {
-    f_type = static_cast<int>(type);    // FIXME cast
+    f_type = type;
 }
